requirements.txt
flask
requests
beautifulsoup4
python-dotenv
streamlit
flask-cors
openai
